Qt: Revert back to an all-in-one main window.
Switch back to having everything on the main window similar to GTK+.
It's more cluttered but this way but you don't have to navigate the task
selector to get to the recent files and help links.

Retain the "Capture", "Open", and "Learn" imperative text. Move "Open"
to the top as suggested by Michal.

Change-Id: Iaf7d93f9caf686b5cbeb640e5e5f95ecb46f7f2d
Reviewed-on: https://code.wireshark.org/review/5896
Reviewed-by: Gerald Combs <gerald@wireshark.org>