Fix handling of 3-component OS version numbers.
Get rid of doubled square brackets (what were those for?), and match
dots in the version number with "\.", so we only match dots.

This fixes the handling of version numbers such as 10.10.1.

(A 2-component version number doesn't match, but that's OK - the sed
command doesn't use -n, so that just means it's unchanged by the
command.)

Change-Id: Ie1cc1587864128b6cefe21841d590cec08024775
Reviewed-on: https://code.wireshark.org/review/5916
Reviewed-by: Guy Harris <guy@alum.mit.edu>