gmr1_rach: Use FT_INT24 instead of FT_UINT24 for latitude/longitude
No need to do sign extension manually, the core code now supports
signed integers with bitmask and it will do the sign extension.
Not sure if that was always the code, but it works now.

This also have the advantage that the real value 'seen' by wireshark
when applying filters is the real signed value.

Change-Id: I47e2357e5556160a00110ad63e3342879769a0b9
Signed-off-by: Sylvain Munaut <tnt@246tNt.com>
Reviewed-on: https://code.wireshark.org/review/6288
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>