Temporarily #if 0 out the call to read_keytab_file().
If this fixes the Win64 build, somehow HAVE_KERBEROS is getting defined
when this is built but we're not actually building read_keytab_file in
packet-kerberos.c.

If that *doesn't* fix the Win64 build's failure to find
read_keytab_file(), something Really Weird is going on, because nobody
should be looking for it.

Change-Id: If607e0eeeff854693cf9ce2ea1009d34a20a9992
Reviewed-on: https://code.wireshark.org/review/6259
Reviewed-by: Guy Harris <guy@alum.mit.edu>