Make the individual install_ and uninstall_ functions idempotent.
Move the "do I want this and, if so, is it already installed?" checks
into the install_ functions.

Move the "is this already installed?" checks into the uninstall_
functions, and have them clear the appropriate installed version
variable when finished.

Also, fetch the installed versions early in the main line.

Change-Id: I9e5d8ca48285c6d424a4a2a5f53f4aa449a891d3
Reviewed-on: https://code.wireshark.org/review/6730
Reviewed-by: Guy Harris <guy@alum.mit.edu>