RTPS: Memset entire buffer
Apparently some compilers do sufficient inlining optimization to notice that
there's at least one caller of rtps_util_add_typecode which passes an
indent_level of 0. Such compilers are *also* picky enough to complain about
memset being called with a length of 0, leading to a warning which we turn into
a compile error.

I thought about putting in an "if indent_level > 0" but I figure memsetting the
entire buffer (length >= 1) is simpler and more efficient in the common case.

Change-Id: Ica21ba988eb0c1251e7b4ef2e110336d5ee32837
Reviewed-on: https://code.wireshark.org/review/6766
Reviewed-by: Evan Huus <eapache@gmail.com>