FT_{U}INT64 doesn't work with BASE_CUSTOM.
This is the 1.12 branch, so we do a quick fix:

Get rid of the one place we're using it.

Reject it in tmp_fld_check_assert(), so using it will blow up early in
the initialization process rather than blowing up when dissecting.

In cases that, with the tmp_fld_check_assert() change, "should not
happen", and where we might *not* be executing in the context of a
dissector (filling in the field label can be done lazily, being deferred
to "print" time, and that doesn't happen in the context of a dissector),
use g_assert_not_reached() rather than DISSECTOR_ASSERT_NOT_REACHED() -
the latter throws an assertion that's not caught if we're not doing
dissection, so we crash anyway.

Bug: 10983
Change-Id: Id4cb1b8ed17cab976b8151e41ec6b6b79478214d
Reviewed-on: https://code.wireshark.org/review/7304
Reviewed-by: Guy Harris <guy@alum.mit.edu>
Reviewed-on: https://code.wireshark.org/review/7705
Reviewed-by: Balint Reczey <balint@balintreczey.hu>
Tested-by: Balint Reczey <balint@balintreczey.hu>