RTP player: don't crash when drawing channel graphs for very long calls.
X11 can't handle pixmaps more than 32k pixels wide so don't try to feed it
one.  This avoids crashes with the complaint "BadAlloc (insufficient
resources for operation)".  Instead we simply truncate the graph (and tell the
user).  Do this in a way that shows the user the empty portion of the graph.

Hopefully the Qt port can fix this properly.

Note: the bug report says that MacOS is not affected but it should be.  So
apply this fix for all on-Windows systems.

Bug: 2630
Change-Id: I71e1bd2f9a62792db06ce887e2bbe7a96d110e0a
Reviewed-on: https://code.wireshark.org/review/7464
Petri-Dish: Jeff Morriss <jeff.morriss.ws@gmail.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>
(cherry picked from commit 701be1f052922aa00288a8c28c42ee358587b7fb)
Reviewed-on: https://code.wireshark.org/review/7478
Reviewed-by: Jeff Morriss <jeff.morriss.ws@gmail.com>