Don't read aliases from the personal hosts file.
We only store the last name added (i.e., the last alias) which is the
opposite of what the name resolver does (it returns the first name in the
hosts file).

Just comment the code out for now with a comment explaining why.  Maybe some
day we'll have a reason to store the aliases.

Change-Id: I560ac05868a429d31e9d41f3a34af36c4fca1f6c
Reviewed-on: https://code.wireshark.org/review/7466
Reviewed-by: Anders Broman <a.broman58@gmail.com>