Fix "Save All" for Export Objects.
Fix the loop that goes through all the objects to be saved so that it
looks at all the objects to be saved rather than repeatedly looking at
the first object.

When letting the user select a directory in which to save, restrict them
to looking at directories.

QDir doesn't do what we actually want here - QDir.setCurrent() might do
a chdir() but it doesn't appear to set the QDir's path, so when we use
it to construct pathnames it constructs them based on
wsApp->lastOpenDir() rather than the directory the user selected.  Use
Boring Old GLib's g_build_filename(), as we do in the GTK+ version, for
now.  (It also does something we *don't* want, namely, at least as the
documentation implies, a chdir().)

Bug: 11062
Change-Id: I705292dc2713b9dba615b0d59a4cf8313f357d82
Reviewed-on: https://code.wireshark.org/review/7720
Reviewed-by: Guy Harris <guy@alum.mit.edu>