dfilter-test.py: add OOT support and Python 3 compat
Support WS_BIN_PATH and SOURCE_DIR (modelled after test/config.sh) to
support out-of-tree builds (including cmake). Add Python 3 support and
mention this in the documentation.

Tested with Python 2.7.9, 3.2.6, 3.4.3:

    WS_BIN_PATH=/tmp/wsbuild/run SOURCE_DIR=/tmp/wireshark \
        tools/dfilter-test.py testBytesIPv6

(2.7.9 and 3.4.3 were also tested fully, but some cases seem to fail
even before this patch. 2.5.6 and 2.6.6 do not work because the unittest
module is outdated.)

Change-Id: I13074579f6f74206edb5cd7be8e7a8406de49c56
Reviewed-on: https://code.wireshark.org/review/7793
Reviewed-by: Anders Broman <a.broman58@gmail.com>